Implemeted a whisper feature, cleaned up the help, fixed the buggy lynch code, added some quips to certain commands.
Also added a function that can find a channel from text but only if it exists and the same thing for a user. I needed it for the whisper command but it can also be used as a catch on all channel and user calls to ensure they exist before the code tries to use them.